requirements.txt
alembic==1.13.0
asgiref==3.8.1
asyncpg==0.30.0
beautifulsoup4==4.13.4
bcrypt==4.1.2
boto3==1.40.51
celery==5.5.2
celery-redbeat==2.3.2
deskew==1.5.1
docling-core==2.50.1
exa-py==1.13.1
fastapi==0.100.0
flower==2.0.1
h5py==3.13.0
html2text==2025.4.15
huggingface_hub==0.35.3
img2pdf==0.6.1
Jinja2==3.1.6
langchain==0.3.27
langchain-anthropic==0.3.22
langchain-aws==0.2.35
langchain-cohere==0.4.6
langchain-exa==0.3.1
langchain-google-genai==2.1.12
langchain-groq==0.3.8
langchain-huggingface==0.3.1
langchain-milvus==0.2.1
langchain-openai==0.3.35
langgraph==0.6.10
langgraph-checkpoint-postgres==2.0.25
msal==1.22.0
msgraph_email==1.1.1
Markdown==3.8.2
numpy==1.26.4
openpyxl==3.1.5
pandas==2.2.1
pdf2image==1.17.0
pikepdf==9.7.0
Pillow==10.4.0
psycopg[binary]==3.2.6
psycopg2-binary==2.9.10
psycopg-pool==3.2.6
pyarmor==9.1.6
pydantic==2.11.4
pymilvus==2.5.14
pymilvus.model==0.3.2
PyMuPDF==1.25.5
PyJWT==2.9.0
pytesseract==0.3.13
python-dotenv==1.1.0
python-jose==3.3.0
python-multipart==0.0.7
premailer==3.10.0
rapidocr==3.3.1
requests==2.32.3
redis==5.3.0
SQLAlchemy==2.0.40
tiktoken==0.8.0
tqdm==4.67.1
uvicorn==0.34.2